Finishing touches
This should be the final commit to this project. The ObjInputTest was made into comments as it wasn't clear on how to make it work or if this was even required. Some smaller issues were fixed as well, and comments were added for any other class or file that needed them.